+++ src/main/scala/edu/ie3/simona/main/RunSimonaStandalone.scala
@@ -59,5 +59,4 @@ object RunSimonaStandalone extends RunSimona[SimonaStandaloneSetup] {
         successful
     }
   }
-
 }