+++ src/test/scala/edu/ie3/simona/sim/setup/SimonaSetupSpec.scala
@@ -13,6 +13,7 @@ import edu.ie3.datamodel.models.input.connector.{
 }
 import edu.ie3.simona.agent.EnvironmentRefs
 import edu.ie3.simona.agent.grid.GridAgent
+import edu.ie3.simona.config.SimonaConfig
 import edu.ie3.simona.event.listener.{ResultEventListener, RuntimeEventListener}
 import edu.ie3.simona.event.{ResultEvent, RuntimeEvent}
 import edu.ie3.simona.ontology.messages.SchedulerMessage
@@ -22,18 +23,26 @@ import edu.ie3.simona.scheduler.core.RegularSchedulerCore
 import edu.ie3.simona.sim.SimonaSim
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.simona.test.common.model.grid.SubGridGateMokka
+import edu.ie3.simona.util.ResultFileHierarchy
 import org.apache.pekko.actor.typed.ActorRef
 import org.apache.pekko.actor.typed.scaladsl.ActorContext
 import org.apache.pekko.actor.{ActorRef => ClassicRef}
 
 import java.nio.file.Path
 import java.util.UUID
+import java.util.concurrent.LinkedBlockingQueue
 
 class SimonaSetupSpec extends UnitSpec with SimonaSetup with SubGridGateMokka {
 
   override val args: Array[String] = Array.empty[String]
 
-  override def logOutputDir: Path = throw new NotImplementedError()
+  override val simonaConfig: SimonaConfig = throw new NotImplementedError()
+  override val resultFileHierarchy: ResultFileHierarchy =
+    throw new NotImplementedError()
+  override val runtimeEventQueue: Option[LinkedBlockingQueue[RuntimeEvent]] =
+    throw new NotImplementedError()
+
+  override val logOutputDir: Path = throw new NotImplementedError()
 
   override def runtimeEventListener(
       context: ActorContext[_]
@@ -43,13 +52,15 @@ class SimonaSetupSpec extends UnitSpec with SimonaSetup with SubGridGateMokka {
     )
 
   override def resultEventListener(
-      context: ActorContext[_]
+      context: ActorContext[_],
+      extSimulationData: ExtSimSetupData,
   ): Seq[ActorRef[ResultEventListener.Request]] =
     throw new NotImplementedException("This is a dummy setup")
 
   override def primaryServiceProxy(
       context: ActorContext[_],
       scheduler: ActorRef[SchedulerMessage],
+      extSimSetupData: ExtSimSetupData,
   ): ClassicRef = throw new NotImplementedException("This is a dummy setup")
 
   override def weatherService(
@@ -60,6 +71,7 @@ class SimonaSetupSpec extends UnitSpec with SimonaSetup with SubGridGateMokka {
   override def extSimulations(
       context: ActorContext[_],
       scheduler: ActorRef[SchedulerMessage],
+      extSimPath: Option[Path],
   ): ExtSimSetupData = throw new NotImplementedException(
     "This is a dummy setup"
   )